requirements.txt
amqp==5.3.1
asgiref==3.8.1
async-timeout==5.0.1
beautifulsoup4==4.13.3
billiard==4.2.1
celery==5.4.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==5.1.7
django-celery-results==2.5.1
django-redis==5.4.0
djangorestframework==3.15.2
dotenv==0.9.9
greenlet==3.1.1
idna==3.10
kombu==5.5.0
mysqlclient==2.2.7
prompt_toolkit==3.0.50
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redis==5.2.1
requests==2.32.3
six==1.17.0
soupsieve==2.6
SQLAlchemy==2.0.39
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
